Move `colorama` from `console_config.py` to `play.py`
It's part of the setup process for the game anyway, so best to just
include it as part of `play.play`'s initialization.
